refactor: 상수 성격에 따라 디렉토리 분리
