+++ frontend/src/components/CategoryAddModal/CategoryAddModal.tsx
@@ -15,7 +15,8 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY, VALIDATION_MESSAGE } from '@/constants';
+import { CACHE_KEY } from '@/constants';
+import { VALIDATION_MESSAGE } from '@/constants/message';
 import { PALETTE } from '@/constants/style';
 
 import { getRandomNumber } from '@/utils';
