+++ frontend/src/components/CategoryModifyModal/CategoryModifyModal.tsx
@@ -13,7 +13,8 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY, VALIDATION_MESSAGE } from '@/constants';
+import { CACHE_KEY } from '@/constants';
+import { VALIDATION_MESSAGE } from '@/constants/message';
 
 import categoryApi from '@/api/category';
