+++ frontend/src/components/ScheduleAddModal/ScheduleAddModal.tsx
@@ -16,8 +16,9 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY, VALIDATION_MESSAGE } from '@/constants';
+import { CACHE_KEY } from '@/constants';
 import { DATE_TIME } from '@/constants/date';
+import { VALIDATION_MESSAGE } from '@/constants/message';
 
 import { getDate, getDateTime } from '@/utils/date';
