+++ frontend/src/components/SubscribedCategoryItem/SubscribedCategoryItem.tsx
@@ -8,7 +8,8 @@ import { userState } from '@/recoil/atoms';
 
 import Button from '@/components/@common/Button/Button';
 
-import { CACHE_KEY, CONFIRM_MESSAGE } from '@/constants';
+import { CACHE_KEY } from '@/constants';
+import { CONFIRM_MESSAGE } from '@/constants/message';
 
 import { getISODateString } from '@/utils/date';
