+++ frontend/src/constants/index.ts
@@ -21,14 +21,6 @@ const CACHE_KEY = {
   SUBSCRIPTIONS: 'subscriptions',
 };
 
-const CONFIRM_MESSAGE = {
-  DELETE_CATEGORY: '정말 삭제하시겠습니까?',
-  UNSUBSCRIBE: '구독을 해제하시겠습니까?',
-  LOGOUT: '로그아웃하시겠습니까?',
-};
-
-const DAYS = ['일', '월', '화', '수', '목', '금', '토'];
-
 const SELECTOR_KEY = {
   SIDE_BAR: 'sideBarSelector',
 };
@@ -43,20 +35,4 @@ const PATH = {
   CATEGORY: '/category',
 };
 
-const VALIDATION_MESSAGE = {
-  STRING_LENGTH: (min: number, max: number) => `${min}자 ~ ${max}자로 입력해주세요`,
-  EARLY_END_DATE_TIME: '시작날짜보다 이후로 설정해주세요.',
-};
-
-export {
-  API,
-  API_URL,
-  ATOM_KEY,
-  CACHE_KEY,
-  CONFIRM_MESSAGE,
-  DAYS,
-  SELECTOR_KEY,
-  STORAGE_KEY,
-  PATH,
-  VALIDATION_MESSAGE,
-};
+export { API, API_URL, ATOM_KEY, CACHE_KEY, SELECTOR_KEY, STORAGE_KEY, PATH };
