+++ frontend/src/constants/message.ts
@@ -1,0 +1,12 @@
+const CONFIRM_MESSAGE = {
+  DELETE_CATEGORY: '정말 삭제하시겠습니까?',
+  UNSUBSCRIBE: '구독을 해제하시겠습니까?',
+  LOGOUT: '로그아웃하시겠습니까?',
+};
+
+const VALIDATION_MESSAGE = {
+  STRING_LENGTH: (min: number, max: number) => `${min}자 ~ ${max}자로 입력해주세요`,
+  EARLY_END_DATE_TIME: '시작날짜보다 이후로 설정해주세요.',
+};
+
+export { CONFIRM_MESSAGE, VALIDATION_MESSAGE };
